// apps/frontend2/src/api/models/createCertificationRequest.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCertificationRequestIssuer } from './createCertificationRequestIssuer';

export interface CreateCertificationRequest {
  /** @minLength 1 */
  name: string;
  issuer?: CreateCertificationRequestIssuer;
  date?: null;
}