// apps/frontend2/src/api/models/profileCertificationDTO.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileCertificationDTOIssuer } from './profileCertificationDTOIssuer';
import type { ProfileCertificationDTODateIssued } from './profileCertificationDTODateIssued';

export interface ProfileCertificationDTO {
  id: number;
  name: string;
  issuer: ProfileCertificationDTOIssuer;
  date_issued: ProfileCertificationDTODateIssued;
  created_at: string;
  updated_at: string;
}