// apps/frontend2/src/api/models/updateCertificationRequest.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateCertificationRequestName } from './updateCertificationRequestName';
import type { UpdateCertificationRequestIssuer } from './updateCertificationRequestIssuer';

export interface UpdateCertificationRequest {
  name?: UpdateCertificationRequestName;
  issuer?: UpdateCertificationRequestIssuer;
  date?: null;
}